# iterator: optimize zeroing of iterAlloc in `Iterator.Close`