@nvanbenschoten 2024-12-03
Here's a listing of the top 10 largest stack frames (before this PR), to give an indication of whether this is a problem elsewhere?
```
➜ go build -gcflags -S . 2>&1 | rg 'ABIInternal' | awk '{print $7 " " $5}' | sed 's/-/ /g' | sed 's/\$//g' | sort -nr | head
17008 8 github.com/cockroachdb/pebble.(*Iterator).Close(SB),
16800 8 github.com/cockroachdb/pebble.(*scanInternalIterator).close(SB),
8656 80 github.com/cockroachdb/pebble.(*compactionPickerByScore).pickAuto(SB),
6288 32 github.com/cockroachdb/pebble.(*pickedCompaction).setupInputs(SB),
4336 88 github.com/cockroachdb/pebble.(*compactionPickerByScore).getScores(SB),
4080 8 github.com/cockroachdb/pebble.(*DB).collectTableStats(SB),
3616 24 github.com/cockroachdb/pebble.(*compaction).newInputIters(SB),
3360 24 github.com/cockroachdb/pebble.Open(SB),
3344 152 github.com/cockroachdb/pebble.ingestLoad1(SB),
3296 48 github.com/cockroachdb/pebble.(*scanInternalIterator).constructPointIter(SB),
```

Notice that `(*Iterator).Close` (fixed here) was the largest offender. `(*scanInternalIterator).close` also contains a `*alloc = iterAlloc{...}`, so we might want to make the same change there, though I don't think that's on a hot path.

There isn't much below that which seems to be on a hot path either. Most of the functions are related to compaction.

I'll run a similar audit in crdb.